Decoupled surface from rhi device
