+++ taichi/rhi/device.h
@@ -581,8 +581,6 @@ class TI_DLL_EXPORT GraphicsDevice : public Device {
 
   virtual Stream *get_graphics_stream() = 0;
 
-  virtual std::unique_ptr<Surface> create_surface(
-      const SurfaceConfig &config) = 0;
   // You are not expected to call this directly. If you want to use this image
   // in a taichi kernel, you usually want to create the image via
   // `GfxRuntime::create_image`. `GfxRuntime` is available in `ProgramImpl`
